requirements.txt
# ── Core Framework ──────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

# ── Database ────────────────────────────────
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
asyncpg==0.29.0
alembic==1.13.2

# ── Redis / Queue ───────────────────────────
redis==5.1.1
apscheduler==3.10.4

# ── HTTP / APIs ─────────────────────────────
httpx==0.27.2
aiohttp==3.10.8

# ── Browser Automation ──────────────────────
playwright==1.48.0

# ── AI / LLM ───────────────────────────────
anthropic==0.37.1

# ── ATS Scoring ─────────────────────────────
scikit-learn==1.5.2

# ── File Storage ────────────────────────────
boto3==1.35.29

# ── PDF / LaTeX ─────────────────────────────
pymupdf==1.24.11

# ── Utilities ───────────────────────────────
python-dotenv==1.0.1
pyyaml==6.0.2
jinja2==3.1.4
python-multipart==0.0.12
websockets==13.1
tenacity==9.0.0
loguru==0.7.2

# ── Notifications ───────────────────────────
python-telegram-bot==21.6
